fix a bug - get the maximum array size of branch correct
The size of array.array was determined by leaf.GetLeadCount().GetMaximum().
However, when a tree is actually a chain with multiple files, this returns only
the maximum of the current file in the chain. The maximum was too small unless
the longest array was in the current file.

This commit fix the bug. It finds the maximum in all files in the chain.

This implementation can be slow, for example, if the chain contains many remote
files.